Send a POST request on submitting login form
This commit enables the form in the login page to send a POST
request (using axios) to the /login route

The cookie sent by this route will be stored in the browser
allowing access to protected route in the future. Cookies
sent on the request object is parsed using the cookieParse
middleware

Previously protected routes could only accessed thorough the a
bearer token. Now the protect route will look for the cookie aswell