Implement logout functionality
Pressing the logout button will override the jwt token with an invalid
one. On success the page will be reloaded from the server instead
of the cache.

A bug was fixed in the isLoggedin function was fixed. After this
commit, any failed promises as a result of an invalid token will
not be routed to the global errorhandling middleware.